refactor: rename modules -> store_modules
