+++ src/namespace_maps.ts
@@ -1,8 +1,8 @@
-import { TodoState } from './modules/todo/store/state';
-import { TodoActions } from './modules/todo/store/actions';
-import { TodoGetters } from './modules/todo/store/getters';
-import { todoStore } from './modules/todo/store';
-import { AuthState } from './modules/auth/store/state';
+import { TodoState } from './store_modules/todo/store/state';
+import { TodoActions } from './store_modules/todo/store/actions';
+import { TodoGetters } from './store_modules/todo/store/getters';
+import { todoStore } from './app_modules/todo/store';
+import { AuthState } from './store_modules/auth/store/state';
 
 export const todoModuleName = 'todo_module';
 export const authModuleName = 'auth';
