+++ src/pages/index.vue
@@ -10,9 +10,9 @@
 import { Component, mixins } from 'nuxt-property-decorator';
 import { StoreHelper, StoreHelperMixin, Actions } from '../mixins/store_helper';
 import * as ns from '../namespace_maps';
-import List from '../modules/todo/components/list.vue';
-import { TodoActions } from '../modules/todo/store/actions';
-import { initialStateResolver } from '../modules/todo/store/index';
+import List from '../store_modules/todo/components/list.vue';
+import { TodoActions } from '../store_modules/todo/store/actions';
+import { initialStateResolver } from '../store_modules/todo/store/index';
 import { Store } from 'vuex';
 
 @Component({
